+++ build_cli/CHANGELOG.md
@@ -4,7 +4,8 @@
   allow users to provide an override to `defaultsTo` by including an optional
   `[fieldName]DefaultOverride` param to the generated 
   `_$populate[ClassName]Parser` function.
-- Handle kebab-case enum value names correctly.
+- Handle kebab-case enum value names correctly, including when they are used as
+  default values.
 - Handle allowed values that require escaping.
 - Require Dart `>=2.2.0 <3.0.0`.